Move ocsp.go and verify.go into lib
This reduces a bunch of copy-paste code.
